requirements.txt
Django==4.0.5
django-cloudinary-storage==0.2.1
tensorflow==2.13.0
urllib3==1.26.5
Pillow
django-debug-toolbar==3.2.1
gunicorn==20.1.0
psycopg2-binary==2.9.5
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
django-crispy-forms==2.0
django-jazzmin==2.6.0